+++ elements/src/modules/ngSelectElements.module.ts
@@ -22,11 +22,6 @@ import {NgSelectElementsComponent} from '../components/ngSelectElements.componen
     declarations:
     [
         NgSelectElementsComponent
-    ],
-    entryComponents:
-    [
-        NgSelectElementsComponent,
-        OptionComponent
     ]
 })
 export class NgSelectElementsModule implements DoBootstrap