+++ src/misc/extensions/setReadonly.ts
@@ -2,7 +2,7 @@ import {NgSelectAction} from "../../components/select";
 
 /**
  * Sets NgSelect as readonly, non editable
- * @param readonly Indication whether is select readonly or not, if omitted defaults to true
+ * @param readonly - Indication whether is select readonly or not, if omitted defaults to true
  */
 export function ɵSetReadonly<TValue>(readonly: boolean = true): NgSelectAction<TValue>
 {