+++ src/misc/optionsGatherer/dynamicOptionsGatherer.interface.ts
@@ -7,7 +7,7 @@ export interface GetOptionsCallback<TValue>
 {
     /**
      * Gets array of options based on query
-     * @param query Query for obtaining options, can be searched string or TValue in case of obtaining option for value during initialization
+     * @param query - Query for obtaining options, can be searched string or TValue in case of obtaining option for value during initialization
      */
     (query: string|TValue): Promise<NgSelectOption<TValue>[]>;
 }