+++ src/plugins/valueHandler/valueHandler.interface.ts
@@ -64,14 +64,14 @@ export interface ValueHandler<TValue> extends NgSelectPlugin
 
     /**
      * Sets value for NgSelect
-     * @param value Value to be set
+     * @param value - Value to be set
      */
     setValue(value:TValue|TValue[]): void;
 
     /**
      * Returns first found options or null
-     * @param text Text of option that is being searched
-     * @param exact Indication whether return only option which is exact match
+     * @param text - Text of option that is being searched
+     * @param exact - Indication whether return only option which is exact match
      */
     findAvailableOption(text: string, exact?: boolean): NgSelectOption<TValue>;
 }